Churn Emails
We have a text file which records mail activity from various individuals in an open source project development team. These files are in a standard format for a file containing multiple mail messages. The lines which start with "From " separate the messages and the lines which start with "From:" are part of the messages. For more information about the mbox format, please see this https://en.wikipedia.org/wiki/Mbox